requirements/test.txt
-r base.txt
pytest
pytest-cov
pytest-django
flake8
factory_boy
